requirements/base.txt
django[argon2]==5.1.2
django-constance==4.1.2
django-htmx==1.19.0
django-stubs[compatible-mypy]==5.1.0
python-dotenv==1.0.1
pytest-django==4.9.0
